% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64serialize}
\alias{base64serialize}
\alias{base64unserialize}
\title{Helpers for converting R objects to strings and back}
\usage{
base64serialize(x, compression = "gzip")

base64unserialize(x, compression = "gzip")
}
\arguments{
\item{x}{Object to serialize or deserialize}

\item{compression}{Passed on as \code{type} argument to \code{\link[=memCompress]{memCompress()}} or
\code{\link[=memDecompress]{memDecompress()}}.}
}
\description{
\code{base64serialize()} converts an R object into a string suitable for storing
in an environment variable. Use this function for encoding entire R objects
(such as OAuth tokens) so that they can be used in Travis builds.

\code{base64unserialize()} is the inverse operation to \code{base64serialize()}.
Use this function in your \code{tic.R} to access the R object previously encoded
by \code{base64serialize()}.
}
\examples{
serial <- base64serialize(1:10)
base64unserialize(serial)
}
