% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\docType{data}
\name{ci}
\alias{ci}
\alias{CI}
\title{The current CI environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ci()

CI
}
\description{
Returns an instance of the CI class that describes the CI environment.
The value is retrieved only once and then cached.
}
\section{Methods}{

\describe{
\item{\code{get_branch()}}{The current branch name, empty if tag.}
\item{\code{get_tag()}}{The current tag name, empty if branch.}
\item{\code{is_tag()}}{Branch or tag?}
\item{\code{is_env()}}{Is the given env variable set to the given value?}
\item{\code{has_env()}}{Does the given env variable exist?}
\item{\code{get_slug()}}{The repo slug in the format \code{user/repo} or \code{org/repo}.}
\item{\code{get_build_number()}}{The build number.}
\item{\code{get_build_url()}}{The URL of the build.}
\item{\code{get_commit()}}{The SHA1 of the current commit.}
\item{\code{is_interactive()}}{
Global setup operations shouldn't be run on an interactive CI,
only on unattended CIs where this method returns \code{FALSE}.}
\item{\code{cat_with_color(code)}}{
Colored output targeted to the CI log.
The \code{code} argument can be an unevaluated call to a
\pkg{crayon} function,
the style will be applied even if it normally wouldn't be.
}
}
}

\keyword{datasets}
