% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-readme-rmd.R
\name{do_readme_rmd}
\alias{do_readme_rmd}
\title{Render a R Markdown README and deploy to Github}
\usage{
do_readme_rmd(
  checkout = TRUE,
  path = ".",
  branch = "master",
  remote_url = NULL,
  commit_message = NULL,
  commit_paths = ".",
  ssh_key_name = "id_rsa"
)
}
\arguments{
\item{checkout}{\verb{[flag]}\cr
Check out the current contents of the repository? Defaults to \code{TRUE},
set to \code{FALSE} if the build process relies on existing contents or
if you deploy to a different branch.}

\item{path, branch}{By default, this macro deploys the \code{"master"} branch
of the readme_rmd repository. An alternative option is \code{"gh-pages"}.}

\item{remote_url}{\verb{[string]}\cr
The URL of the remote Git repository to push to, defaults to the
current GitHub repository.}

\item{commit_message}{\verb{[string]}\cr
Commit message to use, defaults to a useful message linking to the CI build
and avoiding recursive CI runs.}

\item{commit_paths}{\verb{[character]}\cr
Restrict the set of directories and/or files added to Git before deploying.
Default: deploy all files.}

\item{ssh_key_name}{\code{string}\cr
The name of the private SSH key which should be used for deployment.}
}
\description{
\code{do_readme_rmd()} renders an R Markdown README and deploys
the rendered README.md file to Github. It adds default steps to the
\code{"before_deploy"} and \code{"deploy"} stages:

\enumerate{
\item \code{\link[=step_setup_ssh]{step_setup_ssh()}} in the \code{"before_deploy"} to setup
the upcoming deployment
\item \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}} in the \code{"before_deploy"} stage
(if \code{deploy} is set),
\item \code{\link[=step_add_to_readme_rmd]{step_add_to_readme_rmd()}} in the \code{"deploy"}
\item \code{\link[=step_do_push_deploy]{step_do_push_deploy()}} in the \code{"deploy"} stage.
}
}
\examples{
dsl_init()

do_readme_rmd()

dsl_get()
}
\seealso{
Other macros: 
\code{\link{do_bookdown}()},
\code{\link{do_drat}()},
\code{\link{do_package_checks}()},
\code{\link{do_pkgdown}()}
}
\concept{macros}
