% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-ssh.R
\name{encode_private_key}
\alias{encode_private_key}
\title{Encode a private RSA key}
\usage{
encode_private_key(key)
}
\arguments{
\item{key}{RSA key, as returned by \code{\link[openssl:rsa_keygen]{openssl::rsa_keygen()}}}
}
\description{
Encodes a private RSA keygen
}
