% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-ssh.R
\name{get_public_key}
\alias{get_public_key}
\title{Get public RSA key}
\usage{
get_public_key(key)
}
\arguments{
\item{key}{RSA key, as returned by \code{\link[openssl:rsa_keygen]{openssl::rsa_keygen()}}}
}
\description{
Gets a public RSA keygen
}
