% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{list_macros}
\alias{list_macros}
\title{List available macros}
\usage{
list_macros()
}
\value{
\link{character}
}
\description{
Lists available macro functions of the \code{tic} package.
}
\seealso{
Other macros: 
\code{\link{do_bookdown}()},
\code{\link{do_drat}()},
\code{\link{do_package_checks}()},
\code{\link{do_pkgdown}()},
\code{\link{do_readme_rmd}()}
}
\concept{macros}
