% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{load_from_file}
\alias{load_from_file}
\title{Process a tic.R file}
\usage{
load_from_file(path = "tic.R", ..., mtime = file.mtime(path))
}
\description{
Return a named list of stages generated by evaluating a file that uses the
tic \link{DSL}.
}
