% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo}
\alias{repo}
\alias{repo_default}
\alias{repo_cloud}
\alias{repo_cran}
\alias{repo_bioc}
\title{Shortcuts for accessing CRAN-like repositories}
\usage{
repo_default()

repo_cloud()

repo_cran()

repo_bioc(base = repo_default())
}
\arguments{
\item{base}{The base repo to use, defaults to \code{repo_default()}.
Pass \code{NULL} to install only from Bioconductor repos.}
}
\description{
These functions can be used as convenient shortcuts
for the \code{repos} argument to e.g. \code{\link[=do_package_checks]{do_package_checks()}} and
\code{\link[=step_install_deps]{step_install_deps()}}.

\code{repo_default()} returns the value of the \code{"repos"} option,
or \code{repo_cloud()} if the option is not set.

\code{repo_cloud()} returns RStudio's CRAN mirror.

\code{repo_cran()} returns the master CRAN repo.

\code{repo_bioc()} returns Bioconductor repos from
\code{\link[remotes:bioc_install_repos]{remotes::bioc_install_repos()}}, in addition to the default repo.
}
