% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_all_stages}
\alias{run_all_stages}
\title{Emulate a CI run locally}
\usage{
run_all_stages(stages = dsl_load())
}
\arguments{
\item{stages}{\code{[named list]}
A named list of \code{TicStage} objects as returned by \code{\link[=dsl_load]{dsl_load()}},
by default loaded from \code{tic.R}.}
}
\description{
Runs predefined \link{stages} similarly to Travis CI and AppVeyor.
The run aborts on error, the \code{after_failure} stage is never run.
}
\details{
The stages are run in the following order:
\enumerate{
\item \code{before_install()}
\item \code{install()}
\item \code{after_install()}
\item \code{before_script()}
\item \code{script()}
\item \code{after_success()}
\item \code{before_deploy()}
\item \code{deploy()}
\item \code{after_deploy()}
\item \code{after_script()}
}
}
