% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_stage}
\alias{run_stage}
\title{Run a stage}
\usage{
run_stage(name, stages = dsl_load())
}
\arguments{
\item{name}{\code{[string]}\cr
The name of the stage to run.}

\item{stages}{\code{[named list]}
A named list of \code{TicStage} objects as returned by \code{\link[=dsl_load]{dsl_load()}},
by default loaded from \code{tic.R}.}
}
\description{
Run the \code{run_all()} method for all defined steps of a stage for which the
\code{check()} method returns \code{TRUE}.
}
\seealso{
\link{TicStep}

Other runners: \code{\link{prepare_all_stages}}
}
\concept{runners}
