% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{stages}
\alias{stages}
\alias{before_install}
\alias{install}
\alias{after_install}
\alias{before_script}
\alias{script}
\alias{after_success}
\alias{after_failure}
\alias{before_deploy}
\alias{deploy}
\alias{after_deploy}
\alias{after_script}
\title{Predefined stages}
\usage{
before_install(stages = load_from_file())

install(stages = load_from_file())

after_install(stages = load_from_file())

before_script(stages = load_from_file())

script(stages = load_from_file())

after_success(stages = load_from_file())

after_failure(stages = load_from_file())

before_deploy(stages = load_from_file())

deploy(stages = load_from_file())

after_deploy(stages = load_from_file())

after_script(stages = load_from_file())
}
\arguments{
\item{stages}{\code{[named list]}
A named list of \code{Stage} objects as returned by \code{\link[=load_from_file]{load_from_file()}},
by default loaded from \code{rci.R}.}
}
\description{
Stages available in both Travis CI and AppVeyor, for which shortcuts
have been defined. All these functions call \code{\link[=run_stage]{run_stage()}} with the
corresponding stage name.
}
