% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-drat.R
\name{step_add_to_drat}
\alias{step_add_to_drat}
\title{Step: Add built package to a drat}
\usage{
step_add_to_drat(repo_slug = NULL, deploy_dev = FALSE)
}
\arguments{
\item{repo_slug}{\verb{[string]}\cr
The name of the drat repository to deploy to in the form \verb{:owner/:repo}.}

\item{deploy_dev}{\verb{[logical]}\cr
Should development versions of packages also be deployed to the drat repo?
By default only "major", "minor" and "patch" releases are build and
deployed.}
}
\description{
Builds a package (binary on OS X or Windows) and inserts it into an existing
\pkg{drat} repository via \code{\link[drat:insertPackage]{drat::insertPackage()}}.
}
\examples{
dsl_init()

get_stage("script") \%>\%
  add_step(step_add_to_drat())

dsl_get()
}
\seealso{
Other steps: 
\code{\link{step_add_to_known_hosts}()},
\code{\link{step_build_pkgdown}()},
\code{\link{step_do_push_deploy}()},
\code{\link{step_hello_world}()},
\code{\link{step_install_pkg}},
\code{\link{step_install_ssh_keys}()},
\code{\link{step_push_deploy}()},
\code{\link{step_run_code}()},
\code{\link{step_session_info}()},
\code{\link{step_setup_push_deploy}()},
\code{\link{step_setup_ssh}()},
\code{\link{step_test_ssh}()},
\code{\link{step_write_text_file}()}
}
\concept{steps}
