% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-git.R
\name{step_add_to_known_hosts}
\alias{step_add_to_known_hosts}
\title{Step: Add to known hosts}
\usage{
step_add_to_known_hosts(host = "github.com")
}
\arguments{
\item{host}{\code{[string]}\cr
The host name to add to the \code{known_hosts} file, default: \code{github.com}.}
}
\description{
Adds a host name to the \code{~/.ssh/known_hosts} file to allow subsequent
SSH access.
Requires \code{ssh-keyscan} on the system \code{PATH}.
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
