% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-bookdown.R
\name{step_build_bookdown}
\alias{step_build_bookdown}
\title{Step: Build a bookdown book}
\usage{
step_build_bookdown(...)
}
\arguments{
\item{...}{See \link[bookdown:render_book]{bookdown::render_book}.}
}
\description{
Build a bookdown book using \code{\link[bookdown:render_book]{bookdown::render_book()}}.
}
\examples{
dsl_init()

get_stage("script") \%>\%
  add_step(step_build_bookdown("."))

dsl_get()
}
