% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-pkgdown.R
\name{step_build_pkgdown}
\alias{step_build_pkgdown}
\title{Step: Build pkgdown documentation}
\usage{
step_build_pkgdown(...)
}
\arguments{
\item{...}{Passed on to \code{pkgdown::build_site()}}
}
\description{
Builds package documentation with the \pkg{pkgdown} package.
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_pkg}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{steps}
