% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-git.R
\name{step_do_push_deploy}
\alias{step_do_push_deploy}
\title{Step: Perform push deploy}
\usage{
step_do_push_deploy(...)
}
\arguments{
\item{commit_message}{\code{[string]}\cr
Commit message to use, defaults to a useful message linking to the CI build
and avoiding recursive CI runs.}
}
\description{
Commits and pushes to a repo prepared by \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}}.
}
\seealso{
Other deploy steps: \code{\link{step_push_deploy}},
  \code{\link{step_setup_push_deploy}}

Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
