% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-base.R
\name{step_hello_world}
\alias{step_hello_world}
\title{Step: Hello, world!}
\usage{
step_hello_world()
}
\description{
Prints "Hello, world!" to the console.
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
