% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-install.R
\name{step_install_cran}
\alias{step_install_cran}
\title{Step: Install packages from Github}
\usage{
step_install_cran(package = NULL, ...)
}
\arguments{
\item{package}{Package to install}

\item{...}{Passed on to \code{base::install.packages()}}
}
\description{
Install one or more packages from Github
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_github}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{steps}
