% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-install.R
\name{step_install_github}
\alias{step_install_github}
\title{Step: Install packages}
\usage{
step_install_github(repo = NULL, ...)
}
\arguments{
\item{repo}{Package to install in the "user/repo" format.}

\item{...}{Passed on to \code{remotes::install_github()}.}
}
\description{
Install a package from CRAN if its not already installed.
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_cran}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{steps}
