% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-git.R
\name{step_push_deploy}
\alias{step_push_deploy}
\title{Step: Setup and perform push deploy}
\usage{
step_push_deploy(...)
}
\description{
Clones a repo, inits author information, sets up remotes,
commits, and pushes.
Combines \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}} with \code{checkout = FALSE} and
\code{\link[=step_do_push_deploy]{step_do_push_deploy()}}.
}
\seealso{
Other deploy steps: \code{\link{step_do_push_deploy}},
  \code{\link{step_setup_push_deploy}}

Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
