% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-code.R
\name{step_run_code}
\alias{step_run_code}
\title{Step: Run arbitrary R code}
\usage{
step_run_code(call = NULL, prepare_call = NULL)
}
\arguments{
\item{call}{\verb{[call]}\cr
An arbitrary R expression executed during the stage to which this step is
added.
The default is useful if you only pass \code{prepare_call}.}

\item{prepare_call}{\verb{[call]}\cr
An optional arbitrary R expression executed during preparation.}
}
\description{
Captures the expression and executes it when running the step.
An optional preparatory expression can be provided that is executed
during preparation.
If the top-level expression is a qualified function call (of the format
\code{package::fun()}), the package is installed during preparation.
}
\examples{
dsl_init()

get_stage("install") \%>\%
  add_step(step_run_code(update.packages(ask = FALSE)))

# Will install covr from CRAN during preparation:
get_stage("after_success") \%>\%
  add_code_step(covr::codecov())

dsl_get()
}
\seealso{
Other steps: 
\code{\link{step_add_to_drat}()},
\code{\link{step_add_to_known_hosts}()},
\code{\link{step_build_pkgdown}()},
\code{\link{step_do_push_deploy}()},
\code{\link{step_hello_world}()},
\code{\link{step_install_pkg}},
\code{\link{step_install_ssh_keys}()},
\code{\link{step_push_deploy}()},
\code{\link{step_setup_push_deploy}()},
\code{\link{step_setup_ssh}()},
\code{\link{step_test_ssh}()},
\code{\link{step_write_text_file}()}
}
\concept{steps}
