% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-styler.R
\name{step_style_pkg}
\alias{step_style_pkg}
\title{Step: Style package using the \code{styler} package
Styles a R package using \code{styler::style_pkg()}}
\usage{
step_style_pkg(..., pkg = ".", style = styler::tidyverse_style,
  transformers = style(...), filetype = "R",
  exclude_files = "R/RcppExports.R", include_roxygen_examples = TRUE)
}
\arguments{
\item{...}{Ignored, used to enforce naming of arguments.}

\item{pkg}{Path to package.}

\item{style}{A function that creates a style guide to use, by default
\code{tidyverse_style()} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{style_guides()} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{filetype}{Vector of file extensions indicating which file types should
be styled. Case is ignored, and the \code{.} is optional, e.g. \code{c(".R", ".Rmd", ".Rnw")} or \code{c("r", "rmd", "rnw")}.}

\item{exclude_files}{Character vector with paths to files that should be
excluded from styling.}

\item{include_roxygen_examples}{Whether or not to style code in roxygen
examples.}
}
\description{
Step: Style package using the \code{styler} package
Styles a R package using \code{styler::style_pkg()}
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_pkg}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{steps}
