% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-ssh.R
\name{step_test_ssh}
\alias{step_test_ssh}
\title{Step: Test SSH connection}
\usage{
step_test_ssh(
  url = "git@github.com",
  verbose = "",
  private_key_name = "TIC_DEPLOY_KEY"
)
}
\arguments{
\item{url}{\verb{[string]}\cr
URL to establish SSH connection with, by default \code{git@github.com}}

\item{verbose}{\verb{[string]}\cr
Verbosity, by default \code{""}. Use \code{-v} or \code{"-vvv"} for more verbosity.}

\item{private_key_name}{\code{string}\cr
Only needed when deploying from builds on Travis CI or GitHub Actions.
If you have set a custom name for the private key during creation of the
SSH key pair via \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}},
\code{\link[tic:use_ghactions_deploy]{tic::use_ghactions_deploy()}} or \code{\link[=use_tic]{use_tic()}}, pass this name here.}
}
\description{
Establishes an SSH connection.
This step doesn't fail if the connection cannot be established,
but prints verbose output by default.
It is useful for troubleshooting deployment problems.
}
\examples{
dsl_init()

get_stage("script") \%>\%
  add_step(step_test_ssh(verbose = "-vvv"))

dsl_get()
}
\seealso{
Other steps: 
\code{\link{step_add_to_drat}()},
\code{\link{step_add_to_known_hosts}()},
\code{\link{step_build_pkgdown}()},
\code{\link{step_do_push_deploy}()},
\code{\link{step_hello_world}()},
\code{\link{step_install_pkg}},
\code{\link{step_install_ssh_keys}()},
\code{\link{step_push_deploy}()},
\code{\link{step_run_code}()},
\code{\link{step_session_info}()},
\code{\link{step_setup_push_deploy}()},
\code{\link{step_setup_ssh}()},
\code{\link{step_write_text_file}()}
}
\concept{steps}
