% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-write-text-file.R
\name{step_write_text_file}
\alias{step_write_text_file}
\title{Step: Write a text file}
\usage{
step_write_text_file(..., path)
}
\arguments{
\item{...}{\code{[character]}\cr
Contents of the text file.}

\item{path}{\code{[string]}\cr
Path to the new text file.}
}
\description{
Creates a text file with arbitrary contents
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}}, \code{\link{step_test_ssh}}
}
\concept{steps}
