% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic-package.R
\docType{package}
\name{tic-package}
\title{tic: Tasks Integrating Continuously}
\description{
Faciliate deployment tasks for R packages tested by Travis CI,
'AppVeyor', or the CI tool of your choice.
}
\details{
The \code{travis::use_tic()} function prepares a code repository for use with
this package. See \link{DSL} for an overview of \pkg{tic}'s domain-specific
language for defining stages and steps, and \code{\link[=step_hello_world]{step_hello_world()}} and the
links therein for available steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/tic}
  \item Report bugs at \url{https://github.com/ropenscilabs/tic/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}
