% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-base.R
\name{TicStep}
\alias{TicStep}
\title{The base class for all steps}
\description{
Override this class to create a new step.
}
\section{Methods}{

\describe{
\item{\code{run}}{
This method must be overridden, it is called when running the stage
to which a step has been added.
}
\item{\code{prepare}}{
This method is called when preparing the stage
to which a step has been added.
Override this method to install any R packages your step might need,
because this allows them to be cached for subsequential runs.
}
\item{\code{check}}{
This method determines if a step is prepared and run.
Return \code{FALSE} if conditions for running this step are not met.
}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TicStep$new()}}
\item \href{#method-run}{\code{TicStep$run()}}
\item \href{#method-prepare}{\code{TicStep$prepare()}}
\item \href{#method-check}{\code{TicStep$check()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TicStep$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TicStep$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare"></a>}}
\subsection{Method \code{prepare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TicStep$prepare()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TicStep$check()}\if{html}{\out{</div>}}
}

}
}
