% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-yaml.R
\name{update_yml}
\alias{update_yml}
\title{Update tic YAML Templates}
\usage{
update_yml(template_in = NULL, template_out = NULL)
}
\arguments{
\item{template_in}{\verb{[character]}\cr
Path to template which should be updated. By default all standard template
paths of GitHub Actions, Travis CI and Circle CI will be searched and
updated if they exist. Alternatively a full path to a single template can
be passed.}

\item{template_out}{\verb{[character]}\cr
Where the updated template should be written to. This is mainly used for
internal testing purposes and should not be set by the user.
Can only be set}
}
\description{
Updates YAML templates to their
latest versions. Currently only GitHub Actions and Circle CI templates are
supported.
}
\section{Formatting requirements of tic YAML templates}{
 To ensure that
updating of {tic} templates works, ensure the following points:
\itemize{
\item Your template contains the type (e.g. linux-matrix-deploy) and the revision
date in its first two lines.
\item When inserting comments into custom code blocks, only one-line comments are
allowed. Otherwise the update heuristic gets in trouble.
}
}

\examples{
\dontrun{
update_yml("tic.yml")

# custom named templates
update_yml("custom-name.yml")

# full paths
update_yml("~/path/to/repo/.github/workflows/tic.yml")
}
}
\seealso{
yaml_templates
}
