% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-actions.R
\name{use_ghactions_deploy}
\alias{use_ghactions_deploy}
\title{Setup deployment for GitHub Actions}
\usage{
use_ghactions_deploy(
  path = usethis::proj_get(),
  repo = travis::get_repo_slug(remote),
  key_name_private = "TIC_DEPLOY_KEY",
  key_name_public = "Deploy key for GitHub Actions",
  remote = "origin"
)
}
\arguments{
\item{path}{\verb{[string]} \cr
The path to the repository.}

\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{key_name_private}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"TIC_DEPLOY_KEY"} will be used.}

\item{key_name_public}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"Deploy key for GitHub Actions"} will be used.}

\item{remote}{\verb{[string]}\cr
The GitHub remote which should be used. Defaults to "origin".}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Creates a public-private key pair, adds the public key to the GitHub
repository via \code{github_add_key()}, and stores the private key as a "secret"
in the GitHub repo.
}
