% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tic.R
\name{use_tic}
\alias{use_tic}
\title{Initialize CI testing using tic}
\usage{
use_tic(
  wizard = interactive(),
  linux = c("travis", "circle", "none", "all"),
  mac = c("none", "travis", "all"),
  windows = c("none", "appveyor", "all"),
  deploy = c("travis", "circle", "none", "all"),
  matrix = c("none", "travis", "circle", "appveyor", "all"),
  quiet = FALSE
)
}
\arguments{
\item{wizard}{\verb{[flag]}\cr Interactive operation? If \code{TRUE},
a menu will be shown.}

\item{linux, windows, mac}{\verb{[string]}\cr Which CI provider(s) to use to test on
Linux, Windows, and macOS, respectively.}

\item{deploy}{\verb{[string]}\cr Which CI provider(s) to use to
deploy artifacts such as pkgdown documentation.}

\item{matrix}{\verb{[string]}\cr For which CI provider(s) to set up
matrix builds.}

\item{quiet}{\verb{[flag]}\cr Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by
\pkg{tic}.
}
\details{
\enumerate{
\item Query information which CI providers should be used
\item Setup permissions for providers selected for deployment
\item Create YAML files for selected providers
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
}
}
\examples{
# Requires interactive mode
if (FALSE) {
  use_tic()

  # Pre-specified settings favoring Circle CI:
  use_tic(
    linux = "circle",
    mac = "travis",
    windows = "appveyor",
    deploy = "circle",
    matrix = "all"
  )
}
}
