% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_travis_deploy}
\alias{use_travis_deploy}
\title{Set up deployment for Travis CI}
\usage{
use_travis_deploy(travis_repo = NULL, name = "id_rsa", quiet = FALSE)
}
\arguments{
\item{travis_repo}{\code{[string]}\cr
The Travis CI repository to add the private key to, default:
the GitHub repo to which the public deploy key is added.}

\item{name}{\code{[string]}\cr
The private environment variable on Travis CI to write the private key to,
default: \code{"id_rsa"}.  Use this value as \code{name} argument to
\code{\link[=step_setup_ssh]{step_setup_ssh()}} or \code{\link[=step_install_ssh_keys]{step_install_ssh_keys()}}.}

\item{quiet}{\code{[flag]}\cr Less verbose output? Default: \code{FALSE}.}
}
\description{
Creates a public-private key pair,
adds the public key to the GitHub repository via \code{\link[travis:github_add_key]{travis::github_add_key()}},
and stores the private key as an encrypted environment variable in Travis CI
via \code{\link[travis:travis_set_var]{travis::travis_set_var()}}, possibly in a different repository.
See \code{\link[=step_setup_ssh]{step_setup_ssh()}} or \code{\link[=step_install_ssh_keys]{step_install_ssh_keys()}}
for installing such a key during a Travis CI build.
}
