% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-yaml.R
\name{yaml-templates}
\alias{yaml-templates}
\alias{use_travis_yml}
\alias{use_appveyor_yml}
\alias{use_circle_yml}
\title{Use CI YAML templates}
\usage{
use_travis_yml(type)

use_appveyor_yml(type)

use_circle_yml(type)
}
\arguments{
\item{type}{\code{[character]}\cr
Which template to use. The string should be given following the logic
\code{<provider>-<platform>-<action>}. See details for more.}
}
\description{
Installs YAML templates for various CI providers.
}
\section{Type}{

\code{tic} supports a variety of different YAML templates which follow the
\code{<provider>-<platform>-<action>} pattern. The first one is mandatory, the
others are optional.
\itemize{
\item Possible values for \code{<provider>} are \code{travis}, and \code{circle}
\item Possible values for \code{<platform>} are \code{linux}, and \code{macos}, \code{windows}.
\item Possible values for \code{<action>} are \code{matrix} and \code{deploy}.
}

Not every combinations is supported on all CI systems.
For example, for \code{use_appveyor_yaml()} only \code{windows} and \code{windows-matrix} are valid.
}

