% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-yaml.R
\name{yaml_templates}
\alias{yaml_templates}
\alias{use_circle_yml}
\alias{use_ghactions_yml}
\title{Use CI YAML templates}
\usage{
use_circle_yml(type = "linux-deploy", write = TRUE, quiet = FALSE)

use_ghactions_yml(type = "linux-deploy", write = TRUE, quiet = FALSE)
}
\arguments{
\item{type}{\verb{[character]}\cr
Which template to use. The string should be given following the logic
\verb{<platform>-<action>}. See details for more.}

\item{write}{\verb{[logical]}\cr
Whether to write the template to disk (\code{TRUE}) or just return it (\code{FALSE}).}

\item{quiet}{\verb{[logical]}\cr
Whether to print informative messages.}
}
\description{
Installs YAML templates for various CI providers. These functions
are also used within \code{\link[=use_tic]{use_tic()}}.

If you want to update an existing template use \code{\link[=update_yml]{update_yml()}}.
}
\section{pkgdown}{

If \code{type} contains "deploy", {tic} by default also sets the environment
variable \code{BUILD_PKGDOWN=true}. This triggers a call to
\code{pkgdown::build_site()} via the \code{do_pkgdown} macro in \code{tic.R} for the
respective runners.

If a setting  includes "matrix" and builds on multiple R versions, the job
building on R release is chosen to build the pkgdown site.
}

\section{YAML Type}{

\code{tic} supports a variety of different YAML templates which follow the
\verb{<platform>-<action>} pattern. The first one is mandatory, the
others are optional.
\itemize{
\item Possible values for \verb{<platform>} are \code{linux}, and \code{macos}, \code{windows}.
\item Possible values for \verb{<action>} are \code{matrix} and \code{deploy}.
}

Special types are \code{custom} and \code{custom-deploy}. These should be used if the
runner matrix is completely user-defined. This is mainly useful in
\code{\link[=update_yml]{update_yml()}}.

For backward compatibility \code{use_ghactions_yml()} will be default build and
deploy on all platforms.

Here is a list of all available combinations:\tabular{lllll}{
   Provider \tab Operating system \tab Deployment \tab multiple R versions \tab Call \cr
   Circle \tab Linux \tab no \tab no \tab \code{use_circle_yml("linux")} \cr
    \tab Linux \tab yes \tab no \tab \code{use_circle_yml("linux-deploy")} \cr
    \tab Linux \tab no \tab yes \tab \code{use_circle_yml("linux-matrix")} \cr
    \tab Linux \tab no \tab yes \tab \code{use_circle_yml("linux-deploy-matrix")} \cr
   ---------- \tab ------------------------ \tab ---------- \tab ------------------- \tab ------------------------------------------------------- \cr
   GH Actions \tab Linux \tab no \tab no \tab \code{use_ghactions_yml("linux")} \cr
    \tab Linux \tab yes \tab no \tab \code{use_ghactions_yml("linux-deploy")} \cr
    \tab custom \tab no \tab no \tab \code{use_ghactions_yml("custom")} \cr
    \tab custom-deploy \tab yes \tab no \tab \code{use_ghactions_yml("custom-deploy")} \cr
    \tab macOS \tab no \tab no \tab \code{use_ghactions_yml("macos")} \cr
    \tab macOS \tab yes \tab no \tab \code{use_ghactions_yml("macos-deploy")} \cr
    \tab Windows \tab no \tab no \tab \code{use_ghactions_yml("windows")} \cr
    \tab Windows \tab yes \tab no \tab \code{use_ghactions_yml("windows-deploy")} \cr
    \tab Linux + macOS \tab no \tab no \tab \code{use_ghactions_yml("linux-macos")} \cr
    \tab Linux + macOS \tab yes \tab no \tab \code{use_ghactions_yml("linux-macos-deploy")} \cr
    \tab Linux + Windows \tab no \tab no \tab \code{use_ghactions_yml("linux-windows")} \cr
    \tab Linux + Windows \tab yes \tab no \tab \code{use_ghactions_yml("linux-windows-deploy")} \cr
    \tab macOS + Windows \tab no \tab no \tab \code{use_ghactions_yml("macos-windows")} \cr
    \tab macOS + Windows \tab yes \tab no \tab \code{use_ghactions_yml("macos-windows-deploy")} \cr
    \tab Linux + macOS + Windows \tab no \tab no \tab \code{use_ghactions_yml("linux-macos-windows")} \cr
    \tab Linux + macOS + Windows \tab yes \tab no \tab \code{use_ghactions_yml("linux-macos-windows-deploy")} \cr
}
}

