% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{deprecated}
\alias{deprecated}
\alias{add_package_checks}
\title{Deprecated functions}
\usage{
add_package_checks(..., warnings_are_errors = NULL,
  notes_are_errors = NULL, args = c("--no-manual", "--as-cran"),
  build_args = "--force", error_on = "warning",
  repos = getOption("repos"), timeout = Inf)
}
\description{
Deprecated functions of package \code{tic}.

\enumerate{
\item A \code{\link[=step_install_deps]{step_install_deps()}} in the \code{"install"} stage, using the
\code{repos} argument.
\item A \code{\link[=step_rcmdcheck]{step_rcmdcheck()}} in the \code{"script"} stage, using the
\code{warnings_are_errors}, \code{notes_are_errors}, \code{args}, and
\code{build_args} arguments.
\item A call to \code{\link[covr:codecov]{covr::codecov()}} in the \code{"after_success"} stage (only for non-interactive CIs)
}
}
\keyword{internal}
