% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-package-checks.R
\name{Deprecated}
\alias{Deprecated}
\alias{add_package_checks}
\title{Deprecated functions}
\usage{
add_package_checks(..., warnings_are_errors = NULL,
  notes_are_errors = NULL, args = c("--no-manual", "--as-cran"),
  build_args = "--force", error_on = "warning",
  repos = repo_default(), timeout = Inf)
}
\arguments{
\item{...}{Ignored, used to enforce naming of arguments.}

\item{warnings_are_errors}{\code{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{notes_are_errors}{\code{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{c("--no-manual", "--as-cran")}.\cr
Default for Appveyor:
\code{c("--no-manual", "--as-cran", "--no-vignettes", "--no-build-vignettes", "--no-multiarch")}.\cr}

\item{build_args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{"--force"}.\cr
Default for Appveyor: \code{c("--no--build-vignettes", "--force")}.\cr}

\item{error_on}{\code{[character]}\cr
Whether to throw an error on R CMD check failures. Note that the check is
always completed (unless a timeout happens), and the error is only thrown
after completion. If "never", then no errors are thrown. If "error", then
only ERROR failures generate errors. If "warning", then WARNING failures
generate errors as well. If "note", then any check failure generated an
error.}

\item{repos}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{\link[=repo_default]{repo_default()}}.}

\item{timeout}{\code{[numeric]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{Inf}.}
}
\description{
\code{add_package_checks()} has been replaced by \code{\link[=do_package_checks]{do_package_checks()}}.
}
