% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-package-checks.R
\name{do_package_checks}
\alias{do_package_checks}
\title{Add default checks for packages}
\usage{
do_package_checks(
  ...,
  codecov = !ci_is_interactive(),
  warnings_are_errors = NULL,
  notes_are_errors = NULL,
  args = NULL,
  build_args = NULL,
  error_on = "warning",
  repos = repo_default(),
  dependencies = TRUE,
  timeout = Inf,
  check_dir = "check"
)
}
\arguments{
\item{...}{Ignored, used to enforce naming of arguments.}

\item{codecov}{\verb{[flag]}\cr Whether to include a step running
\code{covr::codecov(quiet = FALSE)} (default: only for non-interactive CI,
see \code{\link[=ci_is_interactive]{ci_is_interactive()}}).}

\item{warnings_are_errors, notes_are_errors}{\verb{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{args}{\verb{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr

Default for local runs: \code{c("--no-manual", "--as-cran")}.

Default for Windows:
\code{c("--no-manual", "--as-cran", "--no-vignettes", "--no-build-vignettes", "--no-multiarch")}.

On GitHub Actions option "--no-manual" is always used (appended to custom
user input) because LaTeX is not available and installation is time
consuming and error prone.\cr}

\item{build_args}{\verb{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for local runs: \code{"--force"}.\cr
Default for Windows: \code{c("--no-build-vignettes", "--force")}.\cr}

\item{error_on}{\verb{[character]}\cr
Whether to throw an error on R CMD check failures. Note that the check is
always completed (unless a timeout happens), and the error is only thrown
after completion. If "never", then no errors are thrown. If "error", then
only ERROR failures generate errors. If "warning", then WARNING failures
generate errors as well. If "note", then any check failure generated an
error.}

\item{repos}{\verb{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{\link[=repo_default]{repo_default()}}.}

\item{dependencies}{Dependency types. See
\code{\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies()}} for possible values. Note that
changing this argument from the default might result an installation
failure, e.g. if you set it to \code{FALSE}, packages might not build if
their dependencies are not already installed.}

\item{timeout}{\verb{[numeric]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{Inf}.}

\item{check_dir}{\verb{[character]} \cr Path specifying the directory for R CMD
check. Defaults to \code{"check"} for easy upload of artifacts.}
}
\description{
\code{do_package_checks()} adds default steps related to package checks
to the \code{"before_install"}, \code{"install"}, \code{"script"} and \code{"after_success"}
stages:

This macro is only available for R packages.

\enumerate{
\item \code{\link[=step_install_deps]{step_install_deps()}} in the \code{"install"} stage, using the
\code{repos} argument.
\item \code{\link[=step_session_info]{step_session_info()}} in the \code{"install"} stage.
\item \code{\link[=step_rcmdcheck]{step_rcmdcheck()}} in the \code{"script"} stage, using the
\code{warnings_are_errors}, \code{notes_are_errors}, \code{args}, and
\code{build_args} arguments.
\item A call to \code{\link[covr:codecov]{covr::codecov()}} in the \code{"after_success"} stage
(only if the \code{codecov} flag is set)
}
}
\examples{
dsl_init()

do_package_checks()

dsl_get()
}
\seealso{
Other macros: 
\code{\link{do_blogdown}()},
\code{\link{do_bookdown}()},
\code{\link{do_drat}()},
\code{\link{do_pkgdown}()},
\code{\link{do_readme_rmd}()},
\code{\link{list_macros}()}
}
\concept{macros}
