% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-package-checks.R
\name{do_package_checks}
\alias{do_package_checks}
\title{Add default checks for packages}
\usage{
do_package_checks(..., codecov = !ci_is_interactive(),
  warnings_are_errors = NULL, notes_are_errors = NULL, args = NULL,
  build_args = NULL, error_on = "warning", repos = repo_default(),
  timeout = Inf)
}
\arguments{
\item{...}{Ignored, used to enforce naming of arguments.}

\item{codecov}{\code{[flag]}\cr Whether to include a step running
\code{covr::codecov(quiet = FALSE)} (default: only for non-interactive CI,
see \code{\link[=ci_is_interactive]{ci_is_interactive()}}).}

\item{warnings_are_errors}{\code{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{notes_are_errors}{\code{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{c("--no-manual", "--as-cran")}.\cr
Default for Appveyor:
\code{c("--no-manual", "--as-cran", "--no-vignettes", "--no-build-vignettes", "--no-multiarch")}.\cr}

\item{build_args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{"--force"}.\cr
Default for Appveyor: \code{c("--no--build-vignettes", "--force")}.\cr}

\item{error_on}{\code{[character]}\cr
Whether to throw an error on R CMD check failures. Note that the check is
always completed (unless a timeout happens), and the error is only thrown
after completion. If "never", then no errors are thrown. If "error", then
only ERROR failures generate errors. If "warning", then WARNING failures
generate errors as well. If "note", then any check failure generated an
error.}

\item{repos}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{\link[=repo_default]{repo_default()}}.}

\item{timeout}{\code{[numeric]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{Inf}.}
}
\description{
\code{do_package_checks()} adds default steps related to package checks
to the \code{"before_install"}, \code{"install"}, \code{"script"} and \code{"after_success"}
stages:

\enumerate{
\item \code{\link[=step_install_deps]{step_install_deps()}} in the \code{"install"} stage, using the
\code{repos} argument.
\item \code{\link[=step_rcmdcheck]{step_rcmdcheck()}} in the \code{"script"} stage, using the
\code{warnings_are_errors}, \code{notes_are_errors}, \code{args}, and
\code{build_args} arguments.
\item A call to \code{\link[covr:codecov]{covr::codecov()}} in the \code{"after_success"} stage (only if the \code{codecov} flag is set)
}
}
\examples{
dsl_init()

do_package_checks()

dsl_get()
}
\seealso{
Other macros: \code{\link{do_bookdown}},
  \code{\link{do_pkgdown}}
}
\concept{macros}
