% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-readme-rmd.R
\name{do_readme_rmd}
\alias{do_readme_rmd}
\title{Render a R Markdown README and deploy to Github}
\usage{
do_readme_rmd(
  checkout = TRUE,
  remote_url = NULL,
  commit_message = NULL,
  force = FALSE,
  private_key_name = "TIC_DEPLOY_KEY"
)
}
\arguments{
\item{checkout}{\verb{[flag]}\cr
Check out the current contents of the repository? Defaults to \code{TRUE},
set to \code{FALSE} if the build process relies on existing contents or
if you deploy to a different branch.}

\item{remote_url}{\verb{[string]}\cr
The URL of the remote Git repository to push to, defaults to the
current GitHub repository.}

\item{commit_message}{\verb{[string]}\cr
Commit message to use, defaults to a useful message linking to the CI build
and avoiding recursive CI runs.}

\item{force}{\verb{[logical]}\cr
Add \code{--force} flag to git commands?}

\item{private_key_name}{\code{string}\cr
Only needed when deploying from builds on Travis CI or GitHub Actions.
If you have set a custom name for the private key during creation of the
SSH key pair via \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}},
\code{\link[tic:use_ghactions_deploy]{tic::use_ghactions_deploy()}} or \code{\link[=use_tic]{use_tic()}}, pass this name here.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

\code{do_readme_rmd()} renders an R Markdown README and deploys
the rendered README.md file to Github. It adds default steps to the
\code{"before_deploy"} and \code{"deploy"} stages:

\enumerate{
\item \code{\link[=step_setup_ssh]{step_setup_ssh()}} in the \code{"before_deploy"} to setup
the upcoming deployment
\item \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}} in the \code{"before_deploy"} stage
\item \code{rmarkdown::render()} in the \code{"deploy"} stage
\item \code{\link[=step_do_push_deploy]{step_do_push_deploy()}} in the \code{"deploy"} stage.
}
}
\examples{
\dontrun{
dsl_init()

do_readme_rmd()

dsl_get()
}
}
\seealso{
Other macros: 
\code{\link{do_blogdown}()},
\code{\link{do_bookdown}()},
\code{\link{do_drat}()},
\code{\link{do_package_checks}()},
\code{\link{do_pkgdown}()},
\code{\link{list_macros}()}
}
\concept{macros}
