% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{encode_private_key}
\alias{encode_private_key}
\title{Encode a private RSA key}
\usage{
encode_private_key(key)
}
\arguments{
\item{key}{RSA key, as returned by \code{openssl::rsa_keygen()}}
}
\description{
Extracts the private key from a public-private key pair
generated by \code{\link[openssl:rsa_keygen]{openssl::rsa_keygen()}}, and encodes it in
base64 encoding.
This key can be stored in the \code{id_rsa} private environment variable on Travis CI,
from where \code{\link[=step_install_ssh_keys]{step_install_ssh_keys()}} will pick it up to provide the CI process
with deployment rights.
}
\seealso{
\code{\link[=step_install_ssh_keys]{step_install_ssh_keys()}} \code{\link[=step_test_ssh]{step_test_ssh()}} \code{\link[=step_setup_ssh]{step_setup_ssh()}}
}
\keyword{internal}
