% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{get_public_key}
\alias{get_public_key}
\title{Get public RSA key}
\usage{
get_public_key(key)
}
\arguments{
\item{key}{RSA key, as returned by \code{\link[openssl:rsa_keygen]{openssl::rsa_keygen()}}}
}
\description{
Gets a public RSA keygen
}
\seealso{
\code{\link[usethis:use_ci]{usethis::use_ci()}} \code{\link[usethis:use_travis_deploy]{usethis::use_travis_deploy()}}

\code{\link[=step_install_ssh_keys]{step_install_ssh_keys()}} \code{\link[=step_test_ssh]{step_test_ssh()}} \code{\link[=step_setup_ssh]{step_setup_ssh()}}
}
\keyword{internal}
