% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{load_from_file}
\alias{load_from_file}
\title{Process a tic.R file}
\usage{
load_from_file(path = "tic.R", ..., mtime = file.mtime(path))
}
\arguments{
\item{path}{\code{[string]}\cr
Path to the stage definition file, default: \code{"tic.R"}.}

\item{...}{ignored}

\item{mtime}{\code{[POSIXt]}\cr
For internal use, to detect changes to this file.}
}
\description{
Return a named list of stages generated by \code{\link[=source]{source()}}ing a file that uses the
tic \link{DSL}.
The file is sourced only once per session, unless its modification timestamp
changes.
}
