% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{macro}
\alias{macro}
\title{Macros}
\description{
The \link{DSL} offers a fine-grained interface to the individual stages
of a CI run.
Macros are tic's way of adding several related steps to the relevant
stages.
All macros use the \code{do_} prefix.
}
