% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R, R/macro-package-checks.R
\name{macro}
\alias{macro}
\title{Macros}
\description{
The \link{DSL} offers a fine-grained interface to the individual stages
of a CI run.
Macros are tic's way of adding several related steps to the relevant
stages.
All macros use the \code{do_} prefix.

The \code{\link[=do_package_checks]{do_package_checks()}} macro adds default checks for R packages,
including installation of dependencies and running a test coverage
analysis.
}
