% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{stages}
\alias{stages}
\alias{before_install}
\alias{install}
\alias{after_install}
\alias{before_script}
\alias{script}
\alias{after_success}
\alias{after_failure}
\alias{before_deploy}
\alias{deploy}
\alias{after_deploy}
\alias{after_script}
\title{Predefined stages}
\usage{
before_install(stages = dsl_load())

install(stages = dsl_load())

after_install(stages = dsl_load())

before_script(stages = dsl_load())

script(stages = dsl_load())

after_success(stages = dsl_load())

after_failure(stages = dsl_load())

before_deploy(stages = dsl_load())

deploy(stages = dsl_load())

after_deploy(stages = dsl_load())

after_script(stages = dsl_load())
}
\arguments{
\item{stages}{\code{[named list]}
A named list of \code{TicStage} objects as returned by \code{\link[=dsl_load]{dsl_load()}},
by default loaded from \code{tic.R}.}
}
\description{
Stages available in both Travis CI and AppVeyor, for which shortcuts
have been defined. All these functions call \code{\link[=run_stage]{run_stage()}} with the
corresponding stage name.
}
