% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-drat.R
\name{step_add_to_drat}
\alias{step_add_to_drat}
\title{Step: Add built package to a drat}
\usage{
step_add_to_drat()
}
\description{
Builds a package (binary on OS X or Windows) and inserts it into an existing
\pkg{drat} repository via \code{\link[drat:insertPackage]{drat::insertPackage()}}.
Also knits the \code{index.Rmd} file of the drat if it exists.
}
\seealso{
Other steps: \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_cran}},
  \code{\link{step_install_github}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{steps}
