% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-blogdown.R
\name{step_build_blogdown}
\alias{step_build_blogdown}
\title{Step: Build a Blogdown Site}
\usage{
step_build_blogdown(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[blogdown:build_site]{blogdown::build_site}}
  \describe{
    \item{\code{local}}{Whether to build the website locally to be served via
\code{\link[blogdown]{serve_site}()}.}
    \item{\code{method}}{Different methods to build a website (each with pros and cons).
See Details. The value of this argument will be obtained from the global
option \code{getOption('blogdown.method')} when it is set.}
    \item{\code{run_hugo}}{Whether to run \code{hugo_build()} after R Markdown files are
compiled.}
  }}
}
\description{
Build a Blogdown site using \code{\link[blogdown:build_site]{blogdown::build_site()}}.
}
\examples{
dsl_init()

get_stage("script") \%>\%
  add_step(step_build_blogdown("."))

dsl_get()
}
