% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-bookdown.R
\name{step_build_bookdown}
\alias{step_build_bookdown}
\title{Step: Build a bookdown book}
\usage{
step_build_bookdown(...)
}
\arguments{
\item{...}{Passed on to \code{bookdown::render_book()}}
}
\description{
Build a bookdown book using \code{\link[bookdown:render_book]{bookdown::render_book()}}.
}
