% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-bookdown.R
\name{step_build_bookdown}
\alias{step_build_bookdown}
\title{Step: Build a bookdown book}
\usage{
step_build_bookdown(...)
}
\arguments{
\item{...}{Arguments passed on to \code{bookdown::render_book}
\describe{
  \item{input}{An input filename (or multiple filenames). If \code{preview =
TRUE}, only files specified in this argument are rendered, otherwise all R
Markdown files specified by the book are rendered.}
  \item{output_format}{Arguments to be passed to
\code{rmarkdown::\link[rmarkdown]{render}()}. For \code{preview_chapter()},
\code{...} is passed to \code{render_book()}.}
  \item{clean}{Arguments to be passed to
\code{rmarkdown::\link[rmarkdown]{render}()}. For \code{preview_chapter()},
\code{...} is passed to \code{render_book()}.}
  \item{envir}{Arguments to be passed to
\code{rmarkdown::\link[rmarkdown]{render}()}. For \code{preview_chapter()},
\code{...} is passed to \code{render_book()}.}
  \item{clean_envir}{Whether to clean up the environment \code{envir} before
rendering the book. By default, the environment is cleaned when rendering
the book in a non-interactive R session.}
  \item{output_dir}{The output directory. If \code{NULL}, a field named
\code{output_dir} in the configuration file \file{_bookdown.yml} will be
used (possibly not specified, either, in which case a directory name
\file{_book} will be used).}
  \item{new_session}{Whether to use new R sessions to compile individual Rmd
files (if not provided, the value of the \code{new_session} option in
\file{_bookdown.yml} is used; if this is also not provided,
\code{new_session = FALSE}).}
  \item{preview}{Whether to render and preview the input files specified by the
\code{input} argument. Previewing a certain chapter may save compilation
time as you actively work on this chapter, but the output may not be
accurate (e.g. cross-references to other chapters will not work).}
  \item{encoding}{Ignored. The character encoding of all input files is
supposed to be UTF-8.}
  \item{config_file}{The book configuration file.}
}}
}
\description{
Build a bookdown book using \code{\link[bookdown:render_book]{bookdown::render_book()}}.
}
