% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-git.R
\name{step_do_push_deploy}
\alias{step_do_push_deploy}
\title{Step: Perform push deploy}
\usage{
step_do_push_deploy(path = ".", commit_message = NULL,
  commit_paths = ".")
}
\arguments{
\item{path}{\code{[string]}\cr
Path to the repository, default \code{"."} which means setting up the current
repository.}

\item{commit_message}{\code{[string]}\cr
Commit message to use, defaults to a useful message linking to the CI build
and avoiding recursive CI runs.}

\item{commit_paths}{\code{[character]}\cr
Restrict the set of directories and/or files added to Git before deploying.
Default: deploy all files.}
}
\description{
Commits and pushes to a repo prepared by \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}}.

Deployment usually requires setting up SSH keys with
\code{\link[=use_tic]{use_tic()}} or \code{\link[=use_travis_deploy]{use_travis_deploy()}}.
}
\details{
It is highly recommended to restrict the set of files
touched by the deployment with the \code{commit_paths} argument:
this step assumes that it can freely overwrite all changes to all files
below \code{commit_paths}, and will not warn in case of conflicts.

To mitigate conflicts race conditions to the greatest extent possible,
the following strategy is used:
\itemize{
\item The changes are committed to the branch
\item Before pushing, new commits are fetched, and the changes are cherry-picked
on top of the new commits
}

If no new commits were pushed after the CI run has started,
this strategy is equivalent to committing and pushing.
In the opposite case, if the remote repo has new commits,
the deployment is safely applied to the current tip.
}
\examples{
dsl_init()

# Deployment only works if a companion step_setup_push_deploy() is added
get_stage("deploy") \%>\%
  add_step(step_setup_push_deploy(path = "docs", branch = "gh-pages")) \%>\%
  add_step(step_build_pkgdown())

if (rlang::is_installed("git2r") && git2r::in_repository()) {
  get_stage("deploy") \%>\%
    add_step(step_do_push_deploy(path = "docs"))
}

dsl_get()
}
\seealso{
Other deploy steps: \code{\link{step_push_deploy}},
  \code{\link{step_setup_push_deploy}}

Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_pkg}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
\concept{deploy steps}
\concept{steps}
