% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-ssh.R
\name{step_install_ssh_keys}
\alias{step_install_ssh_keys}
\title{Step: Install an SSH key}
\usage{
step_install_ssh_keys(name = "id_rsa")
}
\arguments{
\item{name}{\code{[string]}\cr
Name of the environment variable and the target file, default: \code{"id_rsa"}.}
}
\description{
Writes a private SSH key encoded in an environment variable
to a file in \code{~/.ssh}.
Only run in non-interactive settings and if the environment variable
exists and is non-empty.
}
\seealso{
\code{travis::use_travis_deploy()}, \code{\link[travis:use_tic]{travis::use_tic()}}

Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_setup_ssh}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
