% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-git.R
\name{step_install_ssh_keys}
\alias{step_install_ssh_keys}
\title{Step: Install an SSH key}
\usage{
step_install_ssh_keys()
}
\description{
Writes a private SSH key encoded in the \code{id_rsa} environment variable
to \code{~/.ssh/id_rsa}.
Only run in non-interactive settings and if the \code{id_rsa} environment variable
exists and is non-empty.
}
\seealso{
\code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}, \code{\link[travis:use_tic]{travis::use_tic()}}

Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_push_deploy}},
  \code{\link{step_run_code}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
