% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-rcmdcheck.R
\name{step_rcmdcheck}
\alias{step_rcmdcheck}
\title{Step: Check a package}
\usage{
step_rcmdcheck(...)
}
\arguments{
\item{warnings_are_errors}{\code{[flag]}\cr
Should warnings be treated as errors? Default: \code{TRUE}.}

\item{notes_are_errors}{\code{[flag]}\cr
Should notes be treated as errors? Default: \code{FALSE}.}

\item{args}{\code{[character]}\cr
Passed to \code{[rcmdcheck::rcmdcheck()]}, default: \code{"--no-manual"}.}
}
\description{
Check a package using \pkg{rcmdcheck}, which ultimately calls \code{R CMD check}.
}
