% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-rcmdcheck.R
\name{step_rcmdcheck}
\alias{step_rcmdcheck}
\title{Step: Check a package}
\usage{
step_rcmdcheck(args = c("--no-manual", "--as-cran"),
  build_args = "--force", error_on = "warning",
  repos = getOption("repos"), timeout = Inf)
}
\arguments{
\item{args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{c("--no-manual", "--as-cran")}.}

\item{build_args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{"--force"}.}

\item{error_on}{\code{[character]}\cr
Whether to throw an error on R CMD check failures. Note that the check is
always completed (unless a timeout happens), and the error is only thrown
after completion. If "never", then no errors are thrown. If "error", then
only ERROR failures generate errors. If "warning", then WARNING failures
generate errors as well. If "note", then any check failure generated an
error.}

\item{repos}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{getOption("repos")}.}

\item{timeout}{\code{[numeric]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{Inf}.}
}
\description{
Check a package using \pkg{rcmdcheck}, which ultimately calls \code{R CMD check}.
The preparation consists of installing package dependencies
via \code{\link[remotes:install_deps]{remotes::install_deps()}} with \code{dependencies = TRUE},
and updating all packages.
}
