% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-rcmdcheck.R
\name{step_rcmdcheck}
\alias{step_rcmdcheck}
\title{Step: Check a package}
\usage{
step_rcmdcheck(warnings_are_errors = TRUE, notes_are_errors = FALSE,
  args = "--no-manual --as-cran")
}
\arguments{
\item{warnings_are_errors}{\code{[flag]}\cr
Should warnings be treated as errors? Default: \code{TRUE}.}

\item{notes_are_errors}{\code{[flag]}\cr
Should notes be treated as errors? Default: \code{FALSE}.}

\item{args}{\code{[character]}\cr
Passed to \code{[rcmdcheck::rcmdcheck()]} (after splitting at spaces), default:
\code{"--no-manual --as-cran"}.}
}
\description{
Check a package using \pkg{rcmdcheck}, which ultimately calls \code{R CMD check}.
}
