% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-rcmdcheck.R
\name{step_rcmdcheck}
\alias{step_rcmdcheck}
\title{Step: Check a package}
\usage{
step_rcmdcheck(..., warnings_are_errors = NULL,
  notes_are_errors = NULL, args = NULL, build_args = NULL,
  error_on = "warning", repos = repo_default(), timeout = Inf)
}
\arguments{
\item{...}{Ignored, used to enforce naming of arguments.}

\item{warnings_are_errors, notes_are_errors}{\code{[flag]}\cr
Deprecated, use \code{error_on}.}

\item{args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{c("--no-manual", "--as-cran")}.\cr
Default for Appveyor:
\code{c("--no-manual", "--as-cran", "--no-vignettes", "--no-build-vignettes", "--no-multiarch")}.\cr}

\item{build_args}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}.\cr
Default for Travis and local runs: \code{"--force"}.\cr
Default for Appveyor: \code{c("--no--build-vignettes", "--force")}.\cr}

\item{error_on}{\code{[character]}\cr
Whether to throw an error on R CMD check failures. Note that the check is
always completed (unless a timeout happens), and the error is only thrown
after completion. If "never", then no errors are thrown. If "error", then
only ERROR failures generate errors. If "warning", then WARNING failures
generate errors as well. If "note", then any check failure generated an
error.}

\item{repos}{\code{[character]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{\link[=repo_default]{repo_default()}}.}

\item{timeout}{\code{[numeric]}\cr
Passed to \code{rcmdcheck::rcmdcheck()}, default:
\code{Inf}.}
}
\description{
Check a package using \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}},
which ultimately calls \code{R CMD check}.
The preparation consists of installing package dependencies
via \code{\link[remotes:install_deps]{remotes::install_deps()}} with \code{dependencies = TRUE},
and updating all packages.
}
\section{Updating of (dependency) packages}{

Packages shipped with the R-installation will not be updated as they will be
overwritten by the Travis R-installer in each build.
If you want these package to be updated, please add the following
step to your workflow: \code{add_code_step(remotes::update_packages(<pkg>)}.
}

