% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-code.R
\name{step_run_code}
\alias{step_run_code}
\title{Step: Run arbitrary code}
\usage{
step_run_code(call)
}
\arguments{
\item{call}{`\link{call}\cr
An arbitrary expression.}
}
\description{
Captures the expression and executes it when running the step.
If the top-level expression is a qualified function call (of the format
\code{package::fun()}), the package is installed during preparation.
}
\examples{
step_run_code(update.packages(ask = FALSE))

# Will install covr from CRAN during preparation:
step_run_code(covr::codecov())
}
\seealso{
Other steps: \code{\link{step_add_to_drat}},
  \code{\link{step_add_to_known_hosts}},
  \code{\link{step_build_pkgdown}},
  \code{\link{step_do_push_deploy}},
  \code{\link{step_hello_world}},
  \code{\link{step_install_ssh_keys}},
  \code{\link{step_push_deploy}},
  \code{\link{step_setup_push_deploy}},
  \code{\link{step_test_ssh}},
  \code{\link{step_write_text_file}}
}
