% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic-package.R
\docType{package}
\name{tic-package}
\alias{tic}
\alias{tic-package}
\title{tic: Tasks Integrating Continuously}
\description{
Faciliate deployment tasks for R packages tested by Travis CI,
'AppVeyor', or the CI tool of your choice.
}
\details{
Use \code{\link{github_repo}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link{github_info}}
function queries basic information on your repository.
The \code{\link{travis_set_var}} and \code{\link{github_add_key}} functions
interact with the GitHub and Travis configuration of your repository,
and ask for permission the first time you call them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/tic}
  \item Report bugs at \url{https://github.com/ropenscilabs/tic/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}

