% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic-package.R
\docType{package}
\name{tic-package}
\alias{tic}
\alias{tic-package}
\title{tic: Tools for tic push stuff}
\description{
More about what it does (maybe more than one line). Stuff!
}
\details{
Use \code{\link{github_repo}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link{github_info}}
function queries basic information on your repository.
The \code{\link{travis_set_var}} and \code{\link{github_add_key}} functions
interact with the GitHub and Travis configuration of your repository,
and ask for permission the first time you call them.
}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}

