% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic-package.R
\docType{package}
\name{tic-package}
\alias{tic-package}
\title{tic: Tasks Integrating Continuously: CI-Agnostic Workflow Definitions}
\description{
Provides a way to describe common build and deployment workflows for R-based
    projects: packages, websites (e.g. blogdown, pkgdown), or data processing
    (e.g. research compendia).  The recipe is described independent of the
    continuous integration tool used for processing the workflow (e.g.
    'Travis CI' or 'AppVeyor').
}
\details{
The \code{\link[=use_tic]{use_tic()}} function prepares a code repository for use with
this package.
See \link{DSL} for an overview of \pkg{tic}'s domain-specific
language for defining stages and steps,
\code{\link[=step_hello_world]{step_hello_world()}} and the links therein for available steps,
and \link{macro} for an overview over the available macros that bundle
several steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/tic}
  \item Report bugs at \url{https://github.com/ropenscilabs/tic/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (0000-0002-1416-3412)

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)
}

Other contributors:
\itemize{
  \item rOpenSci [funder]
}

}
