% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps-base.R
\docType{data}
\name{TicStep}
\alias{TicStep}
\title{The base class for all steps}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TicStep
}
\description{
Override this class to create a new step.
}
\section{Methods}{

\describe{
\item{\code{run}}{
This method must be overridden, it is called when running the stage
to which a step has been added.
}
\item{\code{prepare}}{
This method is called when preparing the stage
to which a step has been added.
Override this method to install any R packages your step might need,
because this allows them to be cached for subsequential runs.
}
\item{\code{check}}{
This method determines if a step is prepared and run.
Return \code{FALSE} if conditions for running this step are not met.
}
}
}

\keyword{datasets}
