% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-yaml.R
\name{update_yml}
\alias{update_yml}
\title{Update tic YAML Templates}
\usage{
update_yml(
  template_in = "inst/test-helpers/test-update-yaml.yml",
  template_out = "inst/test-helpers/test-update-yaml-updated.yml"
)
}
\arguments{
\item{`[character]`\cr}{Where the updated template should be written to. This is mainly used for
internal testing purposes and should not be set by the user.}
}
\description{
Updates the YAML tmps (GitHub Actions, Travis, Circle) to their
latest versions. (For GHA user changes are preserved)
}
