% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-actions.R
\name{use_ghactions_deploy}
\alias{use_ghactions_deploy}
\title{Setup deployment for Travis CI}
\usage{
use_ghactions_deploy(
  path = usethis::proj_get(),
  repo = travis::get_repo_slug(remote),
  key_name_private = "TIC_DEPLOY_KEY",
  key_name_public = "Deploy key for Github Actions",
  remote = "origin",
  quiet = FALSE
)
}
\arguments{
\item{path}{\verb{[string]} \cr
The path to the repository.}

\item{key_name_private}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"TIC_DEPLOY_KEY"} will be used.}

\item{key_name_public}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"Deploy key for Github Actions"} will be used.}
}
\description{
Creates a public-private key pair,
adds the public key to the GitHub repository via \code{github_add_key()},
and stores the private key as an encrypted environment variable in Travis CI
via \code{\link[=travis_set_var]{travis_set_var()}},
possibly in a different repository.
The \pkg{tic} companion package contains facilities for installing such a key
during a Travis CI build.
}
