% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tic.R
\name{use_tic}
\alias{use_tic}
\title{Initialize CI testing using tic}
\usage{
use_tic(
  wizard = interactive(),
  linux = "ghactions",
  mac = "ghactions",
  windows = "ghactions",
  deploy = "ghactions",
  matrix = "none",
  private_key_name = "TIC_DEPLOY_KEY",
  quiet = FALSE
)
}
\arguments{
\item{wizard}{\verb{[flag]}\cr Interactive operation? If \code{TRUE}, a menu will be
shown.}

\item{linux}{\verb{[string]}\cr Which CI provider(s) to use to test on Linux.
Possible options are \code{"circle"}, \code{"ghactions"}, \code{"none"}/\code{NULL}
and \code{"all"}.}

\item{mac}{\verb{[string]}\cr Which CI provider(s) to use to test on macOS
Possible options are \code{"none"}/\code{NULL} and \code{"ghactions"}.}

\item{windows}{\verb{[string]}\cr Which CI provider(s) to use to test on Windows
Possible options are \code{"none"}/\code{NULL}, \code{"appveyor"} and \code{"ghactions"}.}

\item{deploy}{\verb{[string]}\cr Which CI provider(s) to use to deploy artifacts
such as pkgdown documentation. Possible options are "circle"\verb{, }"ghactions"\verb{, }"none"\code{/}NULL\code{and}"all"`.}

\item{matrix}{\verb{[string]}\cr For which CI provider(s) to set up matrix builds.
Possible options are  \code{"circle"}, \code{"ghactions"}, \code{"none"}/\code{NULL}
and \code{"all"}.}

\item{private_key_name}{\code{string}\cr
Only needed when deploying from builds on GitHub Actions.
If you have set a custom name for the private key during creation of the
SSH key pair via tic::use_ghactions_deploy()] or \code{\link[=use_tic]{use_tic()}}, pass this
name here.}

\item{quiet}{\verb{[flag]}\cr Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by
\pkg{tic}.
}
\details{
\enumerate{
\item Query information which CI providers should be used
\item Setup permissions for providers selected for deployment
\item Create YAML files for selected providers
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
}
}
\examples{
# Requires interactive mode
if (FALSE) {
  use_tic()

  # Pre-specified settings favoring Circle CI:
  use_tic(
    wizard = FALSE,
    linux = "circle",
    mac = "ghactions",
    windows = "appveyor",
    deploy = "circle",
    matrix = "all"
  )
}
}
