% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tic.R
\name{use_tic}
\alias{use_tic}
\title{Initialize CI testing using tic}
\usage{
use_tic(quiet = FALSE)
}
\arguments{
\item{quiet}{\code{[flag]}\cr Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by
\pkg{tic}.
}
\details{
\enumerate{
\item Query information which CI providers should be used
\item Setup permissions for providers selected for deployment
\item Create YAML files for selected providers
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
}
}
