% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tic.R
\name{use_tic}
\alias{use_tic}
\title{Initialize CI testing using tic}
\usage{
use_tic(
  wizard = interactive(),
  linux = "travis",
  mac = "travis",
  windows = "appveyor",
  deploy = "travis",
  matrix = "none",
  travis_endpoint = ".com",
  travis_key_name_private = NULL,
  travis_key_name_public = NULL,
  quiet = FALSE
)
}
\arguments{
\item{wizard}{\verb{[flag]}\cr Interactive operation? If \code{TRUE}, a menu will be
shown.}

\item{linux}{\verb{[string]}\cr Which CI provider(s) to use to test on Linux.
Possible options are \code{"travis"}, \code{"circle"}, \code{"none"} and \code{"all"}.}

\item{mac}{\verb{[string]}\cr Which CI provider(s) to use to test on macOS
Possible options are \code{"none"}, and \code{"travis"}.}

\item{windows}{\verb{[string]}\cr Which CI provider(s) to use to test on Windows
Possible options are \code{"none"} and \code{"appveyor"}.}

\item{deploy}{\verb{[string]}\cr Which CI provider(s) to use to deploy artifacts
such as pkgdown documentation. Possible options are \code{"travis"}, \code{"circle"},
\code{"none"} and \code{"all"}.}

\item{matrix}{\verb{[string]}\cr For which CI provider(s) to set up matrix builds.
Possible options are \code{"travis"}, \code{"circle"}, \code{"none"} and \code{"all"}.}

\item{travis_endpoint}{\verb{[string]}\cr The Travis CI endpoint to use. Possible
options are \code{".org"} and \code{".com"}. Default is \code{".com"}. See
\code{\link[travis:travis_enable]{travis::travis_enable()}} for more information.}

\item{travis_key_name_private}{See \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}.}

\item{travis_key_name_public}{See \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}.}

\item{quiet}{\verb{[flag]}\cr Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by
\pkg{tic}.
Depending on your choice of CI system, this function requires
properly set up authentication with Travis CI and/or Circle CI.
See \code{vignette("travis", package = "travis")} and
\code{vignette("circle", package = "circle")} for details.
}
\details{
\enumerate{
\item Query information which CI providers should be used
\item Setup permissions for providers selected for deployment
\item Create YAML files for selected providers
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
}
}
\examples{
# Requires interactive mode
if (FALSE) {
  use_tic()

  # Pre-specified settings favoring Circle CI:
  use_tic(
    wizard = FALSE,
    linux = "circle",
    mac = "travis",
    windows = "appveyor",
    deploy = "circle",
    matrix = "all"
  )
}
}
