% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tic.R
\name{use_tic}
\alias{use_tic}
\title{Initialize CI testing using tic}
\usage{
use_tic(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{\code{[string]}\cr
The path to the repo to prepare.}

\item{quiet}{\code{[flag]}\cr
Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by \pkg{tic}.
}
\details{
The preparation consists of the following steps:
\enumerate{
\item If necessary, create a GitHub repository via \code{\link[travis:uses_github]{travis::uses_github()}}
\item Enable Travis via \code{\link[travis:travis_enable]{travis::travis_enable()}}
\item Create a default \code{.travis.yml} file
(overwrite after confirmation in interactive mode only)
\item Create a default \code{appveyor.yml} file
(depending on repo type, overwrite after confirmation
in interactive mode only)
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
\item Enable deployment (if necessary, depending on repo type)
via \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}
\item Create a GitHub PAT and install it on Travis CI via \code{\link[travis:travis_set_pat]{travis::travis_set_pat()}}
}

This function is aimed at supporting the most common use cases.
Users who require more control are advised to manually call the individual
functions.
}
