% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-badge.R
\name{use_tic_badge}
\alias{use_tic_badge}
\title{Add a CI Status Badge to README files}
\usage{
use_tic_badge(provider, branch = NULL, label = "tic")
}
\arguments{
\item{provider}{\code{character(1)}\cr
The CI provider to generate a badge for. Only \code{ghactions} is currently
supported}

\item{branch}{\code{character(1)}\cr
Which branch should the badge represent? Defaults to the default repo
branch.}

\item{label}{\code{character(1)}\cr
Text to use for the badge.}
}
\description{
Adds a CI status badge to \code{README.Rmd} or \code{README.md}. By default the label
is \code{"tic"}.

A custom branch can be specified via argument \code{branch}.
}
\examples{
\dontrun{
use_tic_badge(provider = "ghactions")

# use a different branch
use_tic_badge(provider = "ghactions", branch = "develop")
}
}
