% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{use_travis_vignettes}
\alias{use_travis_vignettes}
\title{Use travis vignettes}
\usage{
use_travis_vignettes(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}
}
\description{
Use travis vignettes
}

