% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-yaml.R
\name{yaml-templates}
\alias{yaml-templates}
\alias{use_travis_yml}
\alias{use_appveyor_yml}
\alias{use_circle_yml}
\title{Use CI YAML templates}
\usage{
use_travis_yml(type)

use_appveyor_yml(type)

use_circle_yml(type)
}
\arguments{
\item{type}{\verb{[character]}\cr
Which template to use. The string should be given following the logic
\verb{<platform>-<action>}. See details for more.}
}
\description{
Installs YAML templates for various CI providers.
}
\section{Type}{

\code{tic} supports a variety of different YAML templates which follow the
\verb{<platform>-<action>} pattern. The first one is mandatory, the
others are optional.
\itemize{
\item Possible values for \verb{<provider>} are \code{travis}, and \code{circle}
\item Possible values for \verb{<platform>} are \code{linux}, and \code{macos}, \code{windows}.
\item Possible values for \verb{<action>} are \code{matrix} and \code{deploy}.
}

Not every combinations is supported on all CI systems.
For example, for \code{use_appveyor_yaml()} only \code{windows} and \code{windows-matrix} are valid.

Here is a list of all available combinations:\tabular{lllll}{
   Provider \tab Operating system \tab Deployment \tab multiple R versions \tab Call \cr
   ---------- \tab ------------------ \tab ------------ \tab --------------------- \tab --------------------------------------------------------------- \cr
   Travis \tab Linux \tab no \tab no \tab \code{use_travis_yml("linux.yml")} \cr
    \tab Linux \tab yes \tab no \tab \code{use_travis_yml("linux-deploy.yml")} \cr
    \tab Linux \tab no \tab yes \tab \code{use_travis_yml("linux-matrix.yml")} \cr
    \tab Linux \tab yes \tab yes \tab \code{use_travis_yml("linux-deploy-matrix.yml")} \cr
    \tab macOS \tab no \tab no \tab \code{use_travis_yml("macOS.yml")} \cr
    \tab macOS \tab yes \tab no \tab \code{use_travis_yml("macos-deploy.yml")} \cr
    \tab macOS \tab no \tab yes \tab \code{use_travis_yml("macos-matrix.yml")} \cr
    \tab macOS \tab yes \tab yes \tab \code{use_travis_yml("macos-deploy-matrix.yml")} \cr
   ---------- \tab ------------------ \tab ------------ \tab --------------------- \tab --------------------------------------------------------------- \cr
   Circle \tab Linux \tab no \tab no \tab \code{use_circle_yml("linux.yml")} \cr
    \tab Linux \tab yes \tab no \tab \code{use_travis_yml("linux-deploy.yml")} \cr
    \tab Linux \tab no \tab yes \tab \code{use_travis_yml("linux-matrix.yml")} \cr
    \tab Linux \tab no \tab yes \tab \code{use_travis_yml("linux-deploy-matrix.yml")} \cr
   ---------- \tab ------------------ \tab ------------ \tab --------------------- \tab --------------------------------------------------------------- \cr
   Appveyor \tab Windows \tab no \tab no \tab \code{use_appveyor_yml("windows.yml")} \cr
    \tab Windows \tab no \tab yes \tab \code{use_travis_yml("windows-matrix.yml")} \cr
}
}

