% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-yaml.R
\name{yaml_templates}
\alias{yaml_templates}
\alias{use_travis_yml}
\alias{use_appveyor_yml}
\alias{use_circle_yml}
\alias{use_ghactions_yml}
\title{Use CI YAML templates}
\usage{
use_travis_yml(type = "linux-macos-deploy-matrix", write = TRUE, quiet = FALSE)

use_appveyor_yml(type = "windows", write = TRUE, quiet = FALSE)

use_circle_yml(type = "linux-matrix-deploy", write = TRUE, quiet = FALSE)

use_ghactions_yml(
  type = "linux-macos-windows-deploy",
  write = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{type}{\verb{[character]}\cr
Which template to use. The string should be given following the logic
\verb{<platform>-<action>}. See details for more.}

\item{write}{\verb{[logical]}\cr
Whether to write the template to disk (\code{TRUE}) or just return it (\code{FALSE}).}

\item{quiet}{\verb{[logical]}\cr
Whether to print informative messages.}
}
\description{
Installs YAML templates for various CI providers. These function
are also used within \code{\link[=use_tic]{use_tic()}}.
}
\section{pkgdown}{

If \code{type} contains "deploy", {tic} by default also sets the environment
variable \code{BUILD_PKGDOWN=true}. This triggers a call to
\code{pkgdown::build_site()} via the \code{do_pkgdown} macro in \code{tic.R} for the
respective runners.

If a setting  includes "matrix" and builds on multiple R versions, the job
building on R release is chosen to build the pkgdown site.
}

\section{YAML Type}{

\code{tic} supports a variety of different YAML templates which follow the
\verb{<platform>-<action>} pattern. The first one is mandatory, the
others are optional.
\itemize{
\item Possible values for \verb{<platform>} are \code{linux}, and \code{macos}, \code{windows}.
\item Possible values for \verb{<action>} are \code{matrix} and \code{deploy}.
}

Not every combinations is supported on all CI systems.
For example, for \code{use_appveyor_yaml()} only \code{windows} and \code{windows-matrix}
are valid.

For backward compatibility \code{use_ghactions_yml()} will be default build and
deploy on all platforms.

Here is a list of all available combinations:
| Provider   | Operating system         | Deployment | multiple R versions | Call                                                    |
| -------    | ----------------         | ---------- | ------------------- | ------------------------------------------------------- |
|----------  |------------------        |------------|---------------------|---------------------------------------------------------|
| Travis     | Linux                    | no         | no                  | \code{use_travis_yml("linux")}                               |
|            | Linux                    | yes        | no                  | \code{use_travis_yml("linux-deploy")}                        |
|            | Linux                    | no         | yes                 | \code{use_travis_yml("linux-matrix")}                        |
|            | Linux                    | yes        | yes                 | \code{use_travis_yml("linux-deploy-matrix")}                 |
|            | macOS                    | no         | no                  | \code{use_travis_yml("macos")}                               |
|            | macOS                    | yes        | no                  | \code{use_travis_yml("macos-deploy")}                        |
|            | macOS                    | no         | yes                 | \code{use_travis_yml("macos-matrix")}                        |
|            | macOS                    | yes        | yes                 | \code{use_travis_yml("macos-deploy-matrix")}                 |
|            | Linux + macOS            | no         | no                  | \code{use_travis_yml("linux-macos")}                         |
|            | Linux + macOS            | yes        | no                  | \code{use_travis_yml("linux-macos-deploy")}                  |
|            | Linux + macOS            | no         | yes                 | \code{use_travis_yml("linux-macos-matrix")}                  |
|            | Linux + macOS            | yes        | yes                 | \code{use_travis_yml("linux-macos-deploy-matrix")}           |
|----------  |------------------        |------------|---------------------|---------------------------------------------------------|
| Circle     | Linux                    | no         | no                  | \code{use_circle_yml("linux")}                               |
|            | Linux                    | yes        | no                  | \code{use_travis_yml("linux-deploy")}                        |
|            | Linux                    | no         | yes                 | \code{use_travis_yml("linux-matrix")}                        |
|            | Linux                    | no         | yes                 | \code{use_travis_yml("linux-deploy-matrix")}                 |
|----------  |------------------        |------------|---------------------|---------------------------------------------------------|
| Appveyor   | Windows                  | no         | no                  | \code{use_appveyor_yml("windows")}                           |
|            | Windows                  | no         | yes                 | \code{use_travis_yml("windows-matrix")}                      |
|----------  |------------------        |------------|---------------------|---------------------------------------------------------|
| GH Actions | Linux                    | no         | no                  | \code{use_ghactions_yml("linux")}                           -|
|            | Linux                    | yes        | no                  | \verb{use_ghactions_yml("linux-deploy)}                      |
|            | macOS                    | no         | no                  | \verb{use_ghactions_yml("macos)}                             |
|            | macOS                    | yes        | no                  | \verb{use_ghactions_yml("macos-deploy)}                      |
|            | Windows                  | no         | no                  | \verb{use_ghactions_yml("windows)}                           |
|            | Windows                  | yes        | no                  | \verb{use_ghactions_yml("windows-deploy)}                    |
|            | Linux + macOS            | no         | no                  | \code{use_ghactions_yml("linux-macos")}                      |
|            | Linux + macOS            | yes        | no                  | \code{use_ghactions_yml("linux-macos-deploy")}               |
|            | Linux + Windows          | no         | no                  | \code{use_ghactions_yml("linux-windows")}                    |
|            | Linux + Windows          | yes        | no                  | \code{use_ghactions_yml("linux-windows-deploy")}             |
|            | macOS + Windows          | no         | no                  | \code{use_ghactions_yml("macos-windows")}                    |
|            | macOS + Windows          | yes        | no                  | \code{use_ghactions_yml("macos-windows-deploy")}             |
|            | Linux + macOS + Windows  | no         | no                  | \code{use_ghactions_yml("linux-macos-windows")}              |
|            | Linux + macOS + Windows  | yes        | no                  | \code{use_ghactions_yml("linux-macos-windows-deploy")}       |
}

